Add files via upload
A Arduino IDE file extension. The implementation of the collection and processing of the data from the waterproof temperature sensor (DS18B20) and a load cell with capacity to measure weights up to 5 kg. The file also has the connection to the Raspberry Pi server that periodically requests temperature and weight sensor data to understand how much water is available at that time and if the temperature is desirable.